+++ CHANGES.md
@@ -24,6 +24,8 @@ CHANGELOG
  - [Improvement] Add a config parameter to disable tldCache (useful for clickhouse type backends)
  - [Improvement] Implement noNullPoints query parameter. Works only with JSON as in graphite-web
  - [Improvement] For *SeriesLists functions, allow to specify default argument (thx to kolobaev@)
+ - [Improvement] Add support for `round` function (thx to kolobaev@)
+
 
 **0.12.5**
  - [Feature] Implement 'highest' function